Fix build failure

+++ src/hotspot/share/gc/shenandoah/shenandoahController.cpp
@@ -23,12 +23,12 @@
  *
  */
 
+#include "gc/shared/allocTracer.hpp"
 #include "gc/shared/gc_globals.hpp"
 #include "gc/shenandoah/shenandoahCollectorPolicy.hpp"
 #include "gc/shenandoah/shenandoahController.hpp"
 #include "gc/shenandoah/shenandoahHeap.hpp"
 #include "gc/shenandoah/shenandoahHeapRegion.inline.hpp"
-#include "gc/shared/allocTracer.hpp"
 
 void ShenandoahController::update_gc_id() {
   _gc_id.add_then_fetch((size_t)1);
@@ -49,7 +49,7 @@ void ShenandoahController::handle_alloc_failure(const ShenandoahAllocRequest& re
     size_t req_byte = req.size() * HeapWordSize;
     log_info(gc)("Failed to allocate %s, " PROPERFMT, req.type_string(), PROPERFMTARGS(req_byte));
     request_gc(cause);
-    AllocTracer::send_allocation_requiring_gc_event(req_byte, get_gc_id());
+    AllocTracer::send_allocation_requiring_gc_event(req_byte, checked_cast<uint>(get_gc_id()));
   }
 
   if (block) {
